backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
groq>=0.11.0
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
curl_cffi==0.13.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
multitasking==0.0.12
numpy==2.2.6
pandas==2.3.2
pandas-ta==0.4.71b0
peewee==4.0.2
platformdirs==4.9.4
protobuf==7.34.1
bcrypt==4.2.1
PyJWT==2.10.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
ta==0.11.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yfinance==1.2.0
psycopg2-binary==2.9.10
